requirements.txt
Django
gunicorn
whitenoise
psycopg2-binary
dj-database-url
djangorestframework